requirements.txt
scrapy
python-telegram-bot 
requests  # Для работы с HashiCorp Vault 